Split the input module up into a directory
